+++ packages/react-core/src/components/Card/__tests__/Card.test.tsx
@@ -35,11 +35,6 @@ test('card with isHoverable applied ', () => {
   expect(view).toMatchSnapshot();
 });
 
-test('card with isHoverableRaised applied ', () => {
-  const view = shallow(<Card isHoverableRaised />).dive();
-  expect(view).toMatchSnapshot();
-});
-
 test('card with isCompact applied ', () => {
   const view = shallow(<Card isCompact />).dive();
   expect(view).toMatchSnapshot();
@@ -64,23 +59,22 @@ test('card with only isSelected applied - not change', () => {
   expect(view.prop('tabIndex')).toBe(undefined);
 });
 
-test('card with isSelectableRaised applied ', () => {
+test('card with isDisabledRaised applied', () => {
+  const view = shallow(<Card isDisabledRaised />).dive();
+  expect(view.prop('className')).toMatch(/non-selectable-raised/);
+});
+
+test('card with isSelectableRaised applied - not change', () => {
   const view = shallow(<Card isSelectableRaised />).dive();
   expect(view.prop('className')).toMatch(/selectable-raised/);
   expect(view.prop('tabIndex')).toBe('0');
 });
 
-test('card with isSelectableRaised and isSelectedRaised applied ', () => {
-  const view = shallow(<Card isSelectableRaised isSelectedRaised />).dive();
+test('card with isSelectableRaised and isSelected applied ', () => {
+  const view = shallow(<Card isSelected isSelectableRaised />).dive();
   expect(view.prop('className')).toMatch(/selectable-raised/);
   expect(view.prop('className')).toMatch(/selected-raised/);
   expect(view.prop('tabIndex')).toBe('0');
-});
-
-test('card with only isSelectedRaised applied - not change', () => {
-  const view = shallow(<Card isSelectedRaised />).dive();
-  expect(view.prop('className')).not.toMatch(/selected-raised/);
-  expect(view.prop('tabIndex')).toBe(undefined);
 });
 
 test('card with isFlat applied', () => {