+++ packages/react-core/src/components/Dropdown/__tests__/DropdownToggle.test.tsx
@@ -107,6 +107,57 @@ describe('state', () => {
     expect(view).toMatchSnapshot();
   });
 
+  test('button variant - primary', () => {
+    const view = mount(
+      <DropdownToggle id="Dropdown Toggle" toggleVariant="primary" parentRef={document.createElement('div')}>
+        Dropdown
+      </DropdownToggle>
+    );
+    expect(view.find('button').prop('className')).toMatch(/m-primary/);
+    expect(view).toMatchSnapshot();
+  });
+
+  test('button variant - secondary', () => {
+    const view = mount(
+      <DropdownToggle id="Dropdown Toggle" toggleVariant="secondary" parentRef={document.createElement('div')}>
+        Dropdown
+      </DropdownToggle>
+    );
+    expect(view.find('button').prop('className')).toMatch(/m-secondary/);
+    expect(view).toMatchSnapshot();
+  });
+
+  test('action split button - renders primary variant', () => {
+    const view = mount(
+      <DropdownToggle
+        id="Dropdown Toggle"
+        toggleVariant="primary"
+        splitButtonItems={[<div key="1">test</div>]}
+        splitButtonVariant="action"
+        parentRef={document.createElement('div')}
+      >
+        Dropdown
+      </DropdownToggle>
+    );
+    expect(view.find('.pf-c-dropdown__toggle').prop('className')).toMatch(/m-primary/);
+    expect(view).toMatchSnapshot();
+  });
+
+  test('split button - does not render primary variant', () => {
+    const view = mount(
+      <DropdownToggle
+        id="Dropdown Toggle"
+        toggleVariant="primary"
+        splitButtonItems={[<div key="0">test</div>]}
+        parentRef={document.createElement('div')}
+      >
+        Dropdown
+      </DropdownToggle>
+    );
+    expect(view.find('.pf-c-dropdown__toggle').prop('className')).not.toMatch(/m-primary/);
+    expect(view).toMatchSnapshot();
+  });
+
   test('class changes', () => {
     const view = mount(
       <DropdownContext.Provider