+++ packages/react-core/src/components/Popover/PopoverFooter.tsx
@@ -2,6 +2,13 @@ import * as React from 'react';
 import styles from '@patternfly/react-styles/css/components/Popover/popover';
 import { css } from '@patternfly/react-styles';
 
+export interface PopoverFooterProps extends React.HTMLProps<HTMLDivElement> {
+  /** Additional classes added to the Popover footer */
+  className?: string;
+  /** Footer node */
+  children: React.ReactNode;
+}
+
 export const PopoverFooter: React.FunctionComponent<PopoverFooterProps> = ({
   children,
   className = '',
@@ -12,10 +19,3 @@ export const PopoverFooter: React.FunctionComponent<PopoverFooterProps> = ({
   </footer>
 );
 PopoverFooter.displayName = 'PopoverFooter';
-
-export interface PopoverFooterProps extends React.HTMLProps<HTMLDivElement> {
-  /** Additional classes added to the Popover Footer */
-  className?: string;
-  /** Footer node */
-  children: React.ReactNode;
-}