+++ packages/react-integration/cypress/integration/card.spec.ts
@@ -13,6 +13,16 @@ describe('Card Demo Test', () => {
     cy.focused().should('not.have.class', 'pf-m-selected');
   });
 
+  it('Verify that selectableRaised card can be selected and unselected with keyboard input', () => {
+    cy.get('#selectableCardRaised').focus();
+    cy.focused().should('have.class', 'pf-m-selectable-raised');
+    cy.focused().should('not.have.class', 'pf-m-selected-raised');
+    cy.focused().type('{enter}');
+    cy.focused().should('have.class', 'pf-m-selected-raised');
+    cy.focused().type('{enter}');
+    cy.focused().should('not.have.class', 'pf-m-selected-raised');
+  });
+
   it('Verify card is expandable', () => {
     cy.get('#expand-card').should('not.have.class', 'pf-m-expanded');
     cy.get('#expand-card .pf-c-card__header').should('have.class', 'pf-m-toggle-right');